// src/features/settings/types/settings.types.ts
export interface AppSettings {

    choir_name: string;

    church_name: string;

    currency: string;

    registration_open: boolean;

}

export interface AppSettings {

    choir_name: string;

    church_name: string;

    currency: string;

    registration_open: boolean;

    monthly_contribution_amount: number;

}